Extracted out functions associated with CollectionTask.UpdateNote
The following code extractions has been made:

* CollectionTask.UpdateNote::task to a function updateNote.
* AbstractFlashCardViewer.mUpdateCardHandler::onPostExecute to onCardUpdated
* CardBrowser.UpdateCardHandler::onPostExecute to onCardUpdated.
